+++ src/core/login/login-selectors.ts
@@ -24,14 +24,15 @@ export function getStashByUsername(
   username: string
 ): LoginStash | undefined {
   const { stashes } = ai.props.state.login
-  for (const stash of stashes) {
-    if (stash.username === username) return stash
+  for (const rootStash of stashes) {
+    if (rootStash.username === username) return rootStash
   }
 }
 
 export function getStashById(ai: ApiInput, loginId: Uint8Array): StashLeaf {
   const { stashes } = ai.props.state.login
-  for (const stashTree of stashes) {
+  for (const rootStash of stashes) {
+    const stashTree = rootStash
     const stash = searchTree(stashTree, stash =>
       verifyData(stash.loginId, loginId)
     )