+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: Recover WIP login stashes if login-server network requests fails.
+
 ## 2.38.1 (2025-12-19)
 
 - changed: Don't recalculate numeric `EdgeTransaction.confirmations` value